+++ src/hooks/useBreadcrumbs.ts
@@ -1,7 +1,7 @@
 "use client";
 
 import { useBreadcrumbs as useBreadcrumbsContext } from "@/context/BreadcrumbContext";
-import { Breadcrumb } from "@/app/admin/_components/Breadcrumbs";
+import { Breadcrumb } from "src/app/admin/(modules)/_components/Breadcrumbs";
 import { useEffect } from "react";
 
 export const useBreadcrumbs = (value: Breadcrumb[]) => {
